Created scorer.js
Separated out the scorer methods from the main.js file. Also updated the index.html file to make the button calculate the results. Still need to add support for adding files to the program, and also our scorer needs to be updated because it does not appear to be working correctly.